Delete
Delete completed along with some tidying up.
